refactor personal info form and profile avatar components
