+++ src/modals/FriendInfoScreen.tsx
@@ -90,6 +90,7 @@ const FriendInfoScreenModal = ({ navigation, route }: Props) => {
               <ProfileAvatar
                 photoUrl={friend.photoUrl}
                 displayName={friend.displayName}
+                size={24}
                 customClassName="h-24 w-24"
               />
               <Text className="px-2 font-bold text-lg">{friend.displayName}</Text>
